Fix some bugs #1

+++ src/controllers/document-controllers.js
@@ -1,8 +1,12 @@
 import Document from '../models/Document.js';
+import User from '../models/User.js';
 
 // Create document controller
 export const createDocumentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const existCode = await Document.findOne({ code: req.body.code });
         if (existCode) return res.status(403).json({ code: 403, message: 'Số ký hiệu không được trùng' });
         const newDocument = new Document(req.body);
@@ -18,6 +22,9 @@ export const createDocumentController = async (req, res) => {
 // Upload file controller
 export const uploadFileController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const documentId = req.params.documentId;
         const files = req.files;
 
@@ -43,6 +50,9 @@ export const uploadFileController = async (req, res) => {
 // Delete file url controller
 export const deleteFileUrlController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const filename = req.body.filename;
         if (!filename) return res.status(404).json({ code: 404, message: 'Không tìm thấy file' });
         const document = await Document.findById(req.params.documentId);
@@ -58,6 +68,9 @@ export const deleteFileUrlController = async (req, res) => {
 // Upload document controller
 export const updateDocumentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const document = await Document.findById(req.params.documentId);
         if (!document) return res.status(404).json({ code: 404, message: 'Không tìm thấy văn bản' });
 
@@ -80,6 +93,9 @@ export const updateDocumentController = async (req, res) => {
 // Change document status controller
 export const changeDocumentStatusController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const documentStatus = req.body.documentStatus;
         const document = await Document.findById(req.params.documentId);
         if (!document) return res.status(404).json({ code: 404, message: 'Không tìm thấy văn bản' });
@@ -103,6 +119,9 @@ export const changeDocumentStatusController = async (req, res) => {
 // Change document current location controller
 export const changeDocumentLocationController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const documentLocation = req.body.documentLocation;
         const document = await Document.findById(req.params.documentId);
         if (!document) return res.status(404).json({ code: 404, message: 'Không tìm thấy văn bản' });
@@ -126,6 +145,9 @@ export const changeDocumentLocationController = async (req, res) => {
 // Delete document controller
 export const deleteDocumentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const document = await Document.findById(req.params.documentId);
         if (!document) return res.status(404).json({ code: 404, message: 'Không tìm thấy văn bản' });
 
@@ -140,6 +162,9 @@ export const deleteDocumentController = async (req, res) => {
 // Delete many documents controller
 export const deleteManyDocumentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const arrayId = req.body.arrayId;
         await Document.deleteMany({ _id: arrayId });
         res.status(200).json({
@@ -155,6 +180,9 @@ export const deleteManyDocumentController = async (req, res) => {
 // Get all document controller
 export const getAllDocumentController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         let { page, limit, documentName, code, type, status, level, sendDate, documentIn } = req.query;
         const queryFilters = {};
 
@@ -212,6 +240,9 @@ export const getAllDocumentController = async (req, res) => {
 // Get document by ID route
 export const getDocumentByIdController = async (req, res) => {
     try {
+        const currentUser = await User.findById(req.user._id);
+        if (currentUser.isActived === false)
+            return res.status(403).json({ code: 403, message: 'Tài khoản tạm thời bị vô hiệu hóa' });
         const document = await Document.findById(req.params.documentId);
         if (!document) return res.status(404).json({ code: 404, message: 'Không tìm thấy văn bản' });
         res.status(200).json({ code: 200, data: document });
